Refactor GeoTIFF Phase 5b: extract _decode.py from _reader.py (#2246)
Mechanical extraction of strip/tile decode orchestration out of
_reader.py into a new _decode.py module. Behavior-neutral.

Moves the transport-independent decode helpers:

- _apply_predictor, _packed_byte_count, _int_nodata_in_range,
  _resolve_masked_fill, _decode_strip_or_tile
- _read_strips, _read_tiles (the local strip/tile readers)
- _apply_orientation, _apply_orientation_with_geo,
  _apply_photometric_miniswhite, _miniswhite_inverted_nodata
- _NATIVE_ORDER, _PARALLEL_DECODE_PIXEL_THRESHOLD

Left in _reader.py: top-level entry points (_read_to_array,
read_to_array), the COG-HTTP fetch+decode paths, the pixel-safety
guards (MAX_PIXELS_DEFAULT, _check_dimensions,
_check_source_dimensions, PixelSafetyLimitError), and the sparse-
layout helpers (_sparse_fill_value, _has_sparse,
_compute_full_image_byte_budget, _ifd_required_extent) which move
with _layout.py in PR-H (#2247).

_reader.py drops from 2201 to 1407 lines; the moved code lands in
_decode.py at 897 lines. The full public/internal import surface
from xrspatial.geotiff._reader is preserved via a back-import
block, matching the pattern PR-E used for _sources.

Two strip-decode tests that monkey-patched
``_reader_mod.ThreadPoolExecutor`` now patch
``concurrent.futures.ThreadPoolExecutor`` instead, tracking the
decode functions to their new home. The dispatch contract under
test (parallel branch engages on multi-strip, sparse-only short-
circuits) is unchanged.

Part of #2211.
